pyproject.toml
[tool.poetry]
name = "candle-demo"
version = "0.1.0"
description = ""
authors = ["Daniel Mesejo <mesejoleon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
letsql = "^0.1.5"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
